create PageIndex and PageCase, extract Wrapper, add routes
