+++ src/components/PageCase/PageCase.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import Header from "../Header/Header";
+import Wrapper from "../Wrapper/Wrapper";
+
+const PageCase = () => {
+  return (
+    <Wrapper>
+      <Header />
+    </Wrapper>
+  );
+};
+
+export default PageCase;
